HIVE-26392 Move StringableMap tests into a dedicated test class
